+++ spring-core/src/main/java/org/springframework/core/metrics/micrometer/package-info.java
@@ -3,7 +3,7 @@
  */
 @NonNullApi
 @NonNullFields
-package org.springframework.core.metrics.observability;
+package org.springframework.core.metrics.micrometer;
 
 import org.springframework.lang.NonNullApi;
 import org.springframework.lang.NonNullFields;
